[APL-1238] Move alias package to core
