+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/alias/entity/AliasOffer.java
@@ -18,7 +18,7 @@
  * Copyright © 2018-2019 Apollo Foundation
  */
 
-package com.apollocurrency.aplwallet.apl.alias.entity;
+package com.apollocurrency.aplwallet.apl.core.alias.entity;
 
 import com.apollocurrency.aplwallet.apl.core.db.DbKey;
 import com.apollocurrency.aplwallet.apl.core.db.model.VersionedDerivedEntity;
