+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/db/TableRegistryInitializer.java
@@ -9,8 +9,8 @@ import com.apollocurrency.aplwallet.apl.core.account.dao.AccountLeaseTable;
 import com.apollocurrency.aplwallet.apl.core.account.dao.AccountLedgerTable;
 import com.apollocurrency.aplwallet.apl.core.account.dao.AccountPropertyTable;
 import com.apollocurrency.aplwallet.apl.core.account.dao.AccountTable;
-import com.apollocurrency.aplwallet.apl.alias.dao.AliasOfferTable;
-import com.apollocurrency.aplwallet.apl.alias.dao.AliasTable;
+import com.apollocurrency.aplwallet.apl.core.alias.dao.AliasOfferTable;
+import com.apollocurrency.aplwallet.apl.core.alias.dao.AliasTable;
 import com.apollocurrency.aplwallet.apl.core.app.Order;
 import com.apollocurrency.aplwallet.apl.core.app.Poll;
 import com.apollocurrency.aplwallet.apl.core.app.Shuffling;
