+++ apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/account/service/AliasServiceImplTest.java
@@ -1,11 +1,11 @@
 package com.apollocurrency.aplwallet.apl.core.account.service;
 
-import com.apollocurrency.aplwallet.apl.alias.converter.IteratorToStreamConverter;
-import com.apollocurrency.aplwallet.apl.alias.service.AliasServiceImpl;
-import com.apollocurrency.aplwallet.apl.alias.dao.AliasOfferTable;
-import com.apollocurrency.aplwallet.apl.alias.dao.AliasTable;
-import com.apollocurrency.aplwallet.apl.alias.entity.Alias;
-import com.apollocurrency.aplwallet.apl.alias.entity.AliasOffer;
+import com.apollocurrency.aplwallet.apl.core.alias.converter.IteratorToStreamConverter;
+import com.apollocurrency.aplwallet.apl.core.alias.service.AliasServiceImpl;
+import com.apollocurrency.aplwallet.apl.core.alias.dao.AliasOfferTable;
+import com.apollocurrency.aplwallet.apl.core.alias.dao.AliasTable;
+import com.apollocurrency.aplwallet.apl.core.alias.entity.Alias;
+import com.apollocurrency.aplwallet.apl.core.alias.entity.AliasOffer;
 import com.apollocurrency.aplwallet.apl.core.app.Blockchain;
 import com.apollocurrency.aplwallet.apl.core.app.Transaction;
 import com.apollocurrency.aplwallet.apl.core.db.DbClause;
