+++ apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/db/EntityDbTableTest.java
@@ -4,7 +4,7 @@
 
 package com.apollocurrency.aplwallet.apl.core.db;
 
-import com.apollocurrency.aplwallet.apl.alias.converter.IteratorToStreamConverter;
+import com.apollocurrency.aplwallet.apl.core.alias.converter.IteratorToStreamConverter;
 import com.apollocurrency.aplwallet.apl.core.app.Block;
 import com.apollocurrency.aplwallet.apl.core.app.Blockchain;
 import com.apollocurrency.aplwallet.apl.core.app.CollectionUtil;
