+++ apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/dgs/dao/DGSPurchaseTableTest.java
@@ -4,7 +4,7 @@
 
 package com.apollocurrency.aplwallet.apl.core.dgs.dao;
 
-import com.apollocurrency.aplwallet.apl.alias.service.AliasService;
+import com.apollocurrency.aplwallet.apl.core.alias.service.AliasService;
 import com.apollocurrency.aplwallet.apl.core.app.Blockchain;
 import com.apollocurrency.aplwallet.apl.core.app.BlockchainImpl;
 import com.apollocurrency.aplwallet.apl.core.app.BlockchainProcessor;
