+++ apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/phasing/dao/PhasingPollVoterTableTest.java
@@ -4,7 +4,7 @@
 
 package com.apollocurrency.aplwallet.apl.core.phasing.dao;
 
-import com.apollocurrency.aplwallet.apl.alias.service.AliasService;
+import com.apollocurrency.aplwallet.apl.core.alias.service.AliasService;
 import com.apollocurrency.aplwallet.apl.core.app.BlockchainImpl;
 import com.apollocurrency.aplwallet.apl.core.app.GlobalSyncImpl;
 import com.apollocurrency.aplwallet.apl.core.app.TimeServiceImpl;
