+++ apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/tagged/dao/TaggedDataTimestampDaoTest.java
@@ -1,6 +1,6 @@
 package com.apollocurrency.aplwallet.apl.core.tagged.dao;
 
-import com.apollocurrency.aplwallet.apl.alias.service.AliasService;
+import com.apollocurrency.aplwallet.apl.core.alias.service.AliasService;
 import com.apollocurrency.aplwallet.apl.core.app.BlockchainImpl;
 import com.apollocurrency.aplwallet.apl.core.app.GlobalSyncImpl;
 import com.apollocurrency.aplwallet.apl.core.app.TimeServiceImpl;
